Sprint_6_AddClass_AnimalFamilyTest
