+++ src/test/java/AnimalFamilyTest.java
@@ -1,0 +1,14 @@
+import com.example.Animal;
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+
+public class AnimalFamilyTest {
+
+    @Test
+    public void testGetFamily() {
+        Animal animal = new Animal();
+        String expectedFamily = "Существует несколько семейств: заячьи, беличьи, мышиные, кошачьи, псовые, медвежьи, куньи";
+        String actualFamily = animal.getFamily();
+        assertEquals(expectedFamily, actualFamily);
+    }
+}
