+++ src/test/java/AnimalTest.java
@@ -50,11 +50,4 @@ public class AnimalTest {
             }
         }
     }
-    @Test
-    public void testGetFamily() {
-        Animal animal = new Animal();
-        String expectedFamily = "Существует несколько семейств: заячьи, беличьи, мышиные, кошачьи, псовые, медвежьи, куньи";
-        String actualFamily = animal.getFamily();
-        assertEquals(expectedFamily, actualFamily);
-    }
 }
